requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
flake8==7.1.1
Flask==2.1.1
fonttools==4.55.3
fpdf==1.7.2
frozenlist==1.5.0
future==1.0.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
kaleido==0.2.1
kiwisolver==1.4.7
line-bot-sdk==1.20.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mccabe==0.7.0
multidict==6.1.0
numpy==2.2.0
openai==0.28.0
packaging==24.2
pillow==11.0.0
plotly==5.24.1
pluggy==1.5.0
propcache==0.2.1
pycodestyle==2.12.1
pydantic==2.10.3
pydantic_core==2.27.1
pyflakes==3.2.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==2.0.3
yarl==1.18.3
